change bootpwn to a loader script
bootpwn was running /bin/bash but that really doesn't make sense since
much of the time that isn't what we want.  it's likely better to run
whatever command we want and not spawn and extra bash.

this requires the apks to be rebuilt to run "bootpwn <command>" but in
the long run this should lead to a cleaner solution.

depending on how cleanup goes this may be changed later to a two-stage
approach where this loader is run outside the chroot, then another
loader is run inside the chroot, and that loader runs the desired
command.  This type of change wouldn't require rebuilding the apks
again.